refactor(api): extract messages routes into separate module
Move message and event handlers from sessions.rs to new messages.rs
file for better separation of concerns. Sessions module now only
handles session CRUD operations.
